Added support for calculator lookups.
Introduced LohnsteuerCalculator and LohnsteuerCalculation abstractions
to enable users to interact with the generated calculations in a generic
way. LohnsteuerCalculationContextAdapter allows users to configure
a LohnsteuerCalculation with data from a generic map.

It is of course still possible to work with the generated implements
directly.

Revised project setup:
Classes are now generated to target/generated-sources/calculators.
The formatter now only formats the generated code.
The src/main/java Folder was added back to the build-path.